requirements.txt
kafka-python==2.0.2
boto3==1.34.34
snowflake-connector-python==3.7.1
dbt-core==1.7.9
dbt-snowflake==1.7.1
requests==2.31.0